test: Add small test for sycall restart.

+++ user/src/syscall.rs
@@ -477,7 +477,7 @@ pub fn sys_fcntl(fd: usize, cmd: i32, arg: i32) -> isize {
     syscall(SYSCALL_FCNTL, [fd, cmd as usize, arg as usize])
 }
 
-pub fn sys_pipe(pipe: &mut [usize]) -> isize {
+pub fn sys_pipe(pipe: &mut [i32]) -> isize {
     syscall(SYSCALL_PIPE, [pipe.as_mut_ptr() as usize, 0, 0])
 }
